chore: refactor module
