+++ bp-module/Makefile
@@ -1,0 +1,12 @@
+bp-objs := netlink.o proto_sock.o module.o
+obj-m += bp.o
+
+KDIR := /lib/modules/$(shell uname -r)/build
+PWD := $(shell pwd)
+
+all:
+	make -C $(KDIR) M=$(PWD) modules
+
+clean:
+	make -C $(KDIR) M=$(PWD) clean
+
