--- bp-sock/Makefile
@@ -1,9 +1,0 @@
-bp-objs:=netlink.o bp-sock.o
-obj-m += bp.o
-
-all:
-	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) modules
-
-clean:
-	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) clean
-
